requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
pillow==11.0.0
pydantic==2.10.1
pydantic_core==2.27.1
sniffio==1.3.1
typing_extensions==4.12.2
